Fix NPE when invoking context menu in Alerts tab
Fix NullPointerException when attempting to invoke the context menu of
Alerts tab using the keyboard when no entry was yet selected, following
excerpt of the exception's stack trace:
[AWT-EventQueue-0] ERROR org.zaproxy.zap.ZAP$UncaughtExceptionLogger  -
 Exception in thread "AWT-EventQueue-0"
 java.lang.NullPointerException
  at o.z.z.extension.alert.AlertPanel$5.keyPressed(AlertPanel.java:485)

the exception happened because there was no entry selected, therefore no
location existed (that is, null) when calculating where to show the pop
up menu.

Change class AlertPanel to set a JPopupMenu to the Alerts tree, which
already takes care to show the pop up menu in the most appropriate
location (when invoked with keyboard and with the mouse).