Prevent GUI hang when adding messages to History
Change ExtensionHistory to add the HTTP messages (HistoryReference)
asynchronously in the EDT to prevent a deadlock between the EDT and
thread(s) adding messages to the History tab. The ExtensionHistory
synchronises on the history table model object when adding the messages
but it would also add the message synchronously in the EDT, so it would
wait for the EDT while holding the lock, if the EDT tried to add a
message in the meantime a deadlock would happen (since it would try to
sync with the history table model object, already locked).