Add lang directory to manifest of ZAP JAR
Change build.xml file to include the "lang" directory in the manifest of
ZAP jar, so that it's automatically included in the class path. The
"lang" directory contains the resource messages required to run ZAP.
In newer Java 9 (EA) versions the application class loader is no longer
an instance of URLClassLoader ("JEP 261: Module System") making ZAP to
fail (skip) to dynamically add the "lang" directory to the class path.

Related to #2602 - Java 9 (9-ea)