Provide the reason why scanners were skipped
Change class HostProcess to allow to specify (and obtain) the reason why
a scanner was skipped, to specify the reason why the scanner was skipped
when it does not target the selected technologies and to log the reason
when the scanner finishes (if available).
Change class ScanProgressItem to set a reason when the scanner is
skipped by user action.
Change ScanProgressActionIcon to show the reason why a scanner was
skipped (in the tool tip text of the skipped icon), also include the
reason when copying the scanners progress to the clip board.

Fix #2415 - Show the reason why an active scanner was skipped