Minor tidy up of ZAP API client generators
Move some of the common code/behaviour to an abstract class and change
the API generators to extend/use that class (also, extract constants for
the API end points and the optional message).
Change API client generators to use try-with-resource statements when
handling/opening the files and to use Path instead of File and Files to
create the file.
For the JavaAPIGenerator also change to not create the map variable, for
the API parameters, if there's none and change to not create unnecessary
null initialisation for the map and the instance variable "api".