Use the ID to save configs of passive scanners
Change how the configurations of the passive scanners are saved by using
its ID instead of the classname. The ID is used to (uniquely) identify a
scanner so it's more suitable than the classname which gives problems
when the passive scanners are promoted or demoted (since it, usually,
requires a change in the package name).
The configurations now use the elements, "id", "level" and "enabled", to
save the ID, alert threshold and enabled state of a scanner, each one in
its own "pscanner" element (which are all under an "pscan" element).
While the options are migrated some will have an element called
"classname" containing the classname of the scanner, which is removed
once the configurations are saved again (then starting to use the ID).

Change class PluginPassiveScanner to save the configurations with the
new structure, also, document and normalise its behaviour.
Change Constant class to convert to the new options when updating from
version 2.5.0.
Change class PassiveScanParam to expose the root element of the
configurations.
Add tests to assert the expected behaviour of the class
PluginPassiveScanner.