Use PopupMenuItemAlert in core classes
Change classes PopupMenuAlertDelete, PopupMenuAlertEdit and
PopupMenuAlertsRefresh to use PopupMenuItemAlert, instead of "manually"
accessing the Alerts tree.
Change class PopupMenuItemAlert to return just one alert selected if the
menu item does not support multiple selected alerts when a middle alert
node (one that shows the alert name) is selected, instead of returning
all child alerts or the number of all child alerts when checking if the
button of the menu item is enabled for the number of selected alerts.
The change allows menu items that do not support multiple selection to
still be used when the middle alerts are selected (for example, the
edit menu item).
Add some JavaDocs to the classes.