Allow to hook ApiImplementor
Change class ExtensionHook to allow to add ApiImplementor, so it can be
automatically added/removed by core code.
Change ExtensionLoader to add/remove the ApiImplementor when the
extensions are initialised/unloaded.
Change core extensions to use the new method (just removes the coupling
with the API class, as the core extensions are never unloaded).