Allow to keep standard dialogue open after saving
Change StandardFieldsDialog to allow to configure whether or not the
dialogue should be automatically hidden after saving. The change allows,
for example, to keep the dialogue open if the saving process was not
successful to give the user a chance to try to fix the cause of the
problem (that was not possible to detect during the validation process).