Reduce session load time with many alerts
Change SiteNode and HistoryReference to use a Set for the alerts, which
already ensures that there are no duplicates (instead of having to
iterate the Lists to check for duplicates, which was leading to too
much time spent checking that). Also, for SiteNode keep track of the
alert with highest risk instead of iterating the collection each time
the string representation is needed.
Change other classes to reduce the calls to obtain the alerts from the
previously mentioned classes, since they return a copy of the internal
collection.