Correct offset calculation in text header views
Extract the calculation of offsets for view to header and header to view
into a class (HttpTextViewUtils) and changed the text views to use it,
reduces code duplication and uses the correct calculations in all cases
(some calculations were already correct).
Add tests to assert the expected behaviour of HttpTextViewUtils.

Fix #2793 - Wrong highlight in combined view with last part of request
header