Exclude brackets in URLs in spider parsers
Change the regular expressions for HTML comments, in SpiderHtmlParser,
and for text, in SpiderTextParser, to exclude brackets (which are not
expected to be (decoded) in the URLs). Also, change the regular
expression of SpiderTextParser to be case insensitive.
Add tests to assert the expected behaviour of SpiderTextParser.
Update test of SpiderHtmlParser.

Fix #2898 - Tweak spider parser to ignore/strip matched parenthesis
around URLs