Fix concurrency issues when publishing ZAP events
Change SimpleEventBus to control the read/write accesses to the
publishers and consumers to prevent concurrency issues. For example,
when a consumer is unregistered while publishing events, which could
lead to exceptions, e.g.:
java.util.ConcurrentModificationException
 at java.util.ArrayList$Itr.checkForComodification(...)
 at java.util.ArrayList$Itr.next(...)
 at o.z.z.eventBus.SimpleEventBus.publishSyncEvent(...)
 at o.z.z.extension.alert.ExtensionAlert.publishAlertEvent(...)
 at o.z.z.extension.alert.ExtensionAlert.alertFound(...)
 at o.z.z.extension.pscan.PassiveScanThread.raiseAlert(...)
(packages reduced/omitted to keep the lines short)

Only one thread (write access) is allowed to manage the publishers and
consumers while multiple threads can publish events (read access), as
long no thread is managing the publishers or the consumers.
Change the classes RegisteredConsumer and RegisteredPublisher to be
static as they don't need to access the state of SimpleEventBus class.