Correct the loading of extensions' enabled state
Change ExtensionFactory to use the ExtensionParam to obtain the enabled
state of the extensions (which uses the new/correct configuration keys).
Change ExtensionParam to allow to query the enabled state of an
extension and change to use a map to keep the enabled states.
Move ExtensionParam to OptionsParam as it needs to be early available
for core code to use (i.e. ExtensionFactory).
Update tests to assert the new behaviour.

Issue introduced in #2245 - Convert options to not use extensions' names
as XML element names