Allow to deprecate ZAP API endpoints
Change the ZAP API to allow to set its endpoints as deprecated (and add
a description why they are).
Change ZAP API UI to show a note when the endpoints are deprecated.
Change JAVA API generator to annotate and add JavaDoc tag to deprecated
endpoints.

Fix #3061 - Allow to deprecate API endpoints