Do not set the "in scope" state in Contexts panel
Change ContextListPanel to not set the "in scope" state to the contexts
as that might (depending on the internal order of the panels) override
the value set by/in ContextGeneralPanel. The ContextListPanel does not
allow to change the "in scope" state so it should not be setting it.

Fix #3100 - Context's in scope change might not be applied