Improve error handling on .pem root CA cert import
Change DynamicSSLPanel to warn/inform if the .pem file does not contain
the required sections (certificate and private key), also, warn if the
sections are not properly base64 encoded.
Change SslCertificateUtils to support the previous changes.
Add some tests to assert the expected behaviour of SslCertificateUtils.

Fix #2526 - CA Certs errors